refactor(*): 将与模型相关的dataset、task和predictor移入model
