+++ ark_nlp/model/ner/biaffine_bert/__init__.py
@@ -1,5 +1,5 @@
-from ark_nlp.dataset import BiaffineNERDataset as Dataset
-from ark_nlp.dataset import BiaffineNERDataset as BiaffineBertNERDataset
+from ark_nlp.model.ner.biaffine_bert.dataset import BiaffineBertNERDataset as Dataset
+from ark_nlp.model.ner.biaffine_bert.dataset import BiaffineBertNERDataset
 
 from ark_nlp.processor.tokenizer.transformer import TokenTokenizer as Tokenizer
 from ark_nlp.processor.tokenizer.transformer import TokenTokenizer as BiaffineBertNERTokenizer
@@ -7,14 +7,14 @@ from ark_nlp.processor.tokenizer.transformer import TokenTokenizer as BiaffineBe
 from ark_nlp.nn import BertConfig as BiaffineBertConfig
 from ark_nlp.nn import BertConfig as ModuleConfig
 
-from ark_nlp.model.ner.biaffine_bert.biaffine_bert import BiaffineBert
-from ark_nlp.model.ner.biaffine_bert.biaffine_bert import BiaffineBert as Module
+from ark_nlp.model.ner.biaffine_bert.module import BiaffineBert
+from ark_nlp.model.ner.biaffine_bert.module import BiaffineBert as Module
 
 from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_model_optimizer
 from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_biaffine_bert_optimizer
 
-from ark_nlp.model.ner.biaffine_bert.biaffine_named_entity_recognition import BiaffineNERTask as Task
-from ark_nlp.model.ner.biaffine_bert.biaffine_named_entity_recognition import BiaffineNERTask as BiaffineBertNERTask
+from ark_nlp.model.ner.biaffine_bert.task import BiaffineBertNERTask as Task
+from ark_nlp.model.ner.biaffine_bert.task import BiaffineBertNERTask
 
-from ark_nlp.factory.predictor import BiaffineNERPredictor as Predictor
-from ark_nlp.factory.predictor import BiaffineNERPredictor as BiaffineBertNERPredictor
+from ark_nlp.model.ner.biaffine_bert.predictor import BiaffineBertNERPredictor as Predictor
+from ark_nlp.model.ner.biaffine_bert.predictor import BiaffineBertNERPredictor
