+++ ark_nlp/model/ner/crf_bert/__init__.py
@@ -1,20 +1,20 @@
-from ark_nlp.dataset import BIONERDataset as Dataset
-from ark_nlp.dataset import BIONERDataset as CrfBertNERDataset
+from ark_nlp.model.ner.crf_bert.dataset import CrfBertNERDataset as Dataset
+from ark_nlp.model.ner.crf_bert.dataset import CrfBertNERDataset
 
 from ark_nlp.processor.tokenizer.transformer import TokenTokenizer as Tokenizer
 from ark_nlp.processor.tokenizer.transformer import TokenTokenizer as CrfBertNERTokenizer
 
-from ark_nlp.nn import BertConfig as CrfBertConfig
 from ark_nlp.nn import BertConfig as ModuleConfig
+from ark_nlp.nn import BertConfig as CrfBertConfig
 
-from ark_nlp.model.ner.crf_bert.crf_bert import CrfBert
-from ark_nlp.model.ner.crf_bert.crf_bert import CrfBert as Module
+from ark_nlp.model.ner.crf_bert.module import CrfBert as Module
+from ark_nlp.model.ner.crf_bert.module import CrfBert
 
 from ark_nlp.factory.optimizer import get_default_crf_bert_optimizer as get_default_model_optimizer
 from ark_nlp.factory.optimizer import get_default_crf_bert_optimizer as get_default_crf_bert_optimizer
 
-from ark_nlp.factory.task import CRFNERTask as Task
-from ark_nlp.factory.task import CRFNERTask as CrfBertNERTask
+from ark_nlp.factory.task import CrfBertNERTask as Task
+from ark_nlp.factory.task import CrfBertNERTask
 
-from ark_nlp.factory.predictor import CRFNERPredictor as Predictor
-from ark_nlp.factory.predictor import CRFNERPredictor as CrfBertNERPredictor
+from ark_nlp.factory.predictor import CrfBertNERPredictor as Predictor
+from ark_nlp.factory.predictor import CrfBertNERPredictor
