+++ ark_nlp/model/ner/crf_bert/dataset.py
@@ -1,0 +1,1 @@
+from ark_nlp.dataset import BIONERDataset as CrfBertNERDataset
