+++ ark_nlp/model/ner/crf_bert/predictor.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2021 DataArk Authors. All Rights Reserved.
+# Copyright (c) 2020 DataArk Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,10 +18,12 @@
 
 import torch
 
+from ark_nlp.factory.utils.conlleval import get_entities
 
-class SpanNERPredictor(object):
+
+class CrfBertNERPredictor(object):
     """
-    span模式的命名实体识别的预测器
+    +CRF模式的字符分类任务的预测器
 
     Args:
         module: 深度学习模型
@@ -33,8 +35,11 @@ class SpanNERPredictor(object):
         self,
         module,
         tokernizer,
-        cat2id
+        cat2id,
+        markup='bio'
     ):
+        self.markup = markup
+
         self.module = module
         self.module.task = 'TokenLevel'
 
@@ -50,19 +55,29 @@ class SpanNERPredictor(object):
         self,
         text
     ):
-        tokens = self.tokenizer.tokenize(text)
-        token_mapping = self.tokenizer.get_token_mapping(text, tokens)
-
-        input_ids = self.tokenizer.sequence_to_ids(tokens)
+        input_ids = self.tokenizer.sequence_to_ids(text)
         input_ids, input_mask, segment_ids = input_ids
 
         features = {
-            'input_ids': input_ids,
-            'attention_mask': input_mask,
-            'token_type_ids': segment_ids,
-        }
+                'input_ids': input_ids,
+                'attention_mask': input_mask,
+                'token_type_ids': segment_ids
+            }
+        return features
 
-        return features, token_mapping
+    def _convert_to_vanilla_ids(
+        self,
+        text
+    ):
+        tokens = self.tokenizer.tokenize(text)
+        length = len(tokens)
+        input_ids = self.tokenizer.sequence_to_ids(tokens)
+
+        features = {
+                'input_ids': input_ids,
+                'length': length if length < self.tokenizer.max_seq_len else self.tokenizer.max_seq_len,
+            }
+        return features
 
     def _get_input_ids(
         self,
@@ -94,36 +109,34 @@ class SpanNERPredictor(object):
             text (:obj:`string`): 输入文本
         """  # noqa: ignore flake8"
 
-        features, token_mapping = self._get_input_ids(text)
+        features = self._get_input_ids(text)
         self.module.eval()
 
         with torch.no_grad():
             inputs = self._get_module_one_sample_inputs(features)
-            start_logits, end_logits = self.module(**inputs)
-            start_scores = torch.argmax(start_logits[0].cpu(), -1).numpy()[1:]
-            end_scores = torch.argmax(end_logits[0].cpu(), -1).numpy()[1:]
+            logit = self.module(**inputs)
+
+        tags = self.module.crf.decode(logit, inputs['attention_mask'])
+        tags = tags.squeeze(0)
+
+        preds = tags.detach().cpu().numpy().tolist()
+        preds = preds[0][1:]
+        preds = preds[:len(text)]
+
+        tags = [self.id2cat[x] for x in preds]
+        label_entities = get_entities(preds, self.id2cat, self.markup)
+
+        entities = set()
+        for entity_ in label_entities:
+            entities.add(text[entity_[1]: entity_[2]+1] + '-' + entity_[0])
 
         entities = []
-        for index_, s_l in enumerate(start_scores):
-            if s_l == 0:
-                continue
-
-            if index_ > token_mapping[-1][-1]:
-                break
-
-            for jndex_, e_l in enumerate(end_scores[index_:]):
-
-                if index_ + jndex_ > token_mapping[-1][-1]:
-                    break
-
-                if s_l == e_l:
-                    entitie_ = {
-                        "start_idx": token_mapping[index_][0],
-                        "end_idx": token_mapping[index_+jndex_][-1],
-                        "type": self.id2cat[s_l],
-                        "entity": text[token_mapping[index_][0]: token_mapping[index_+jndex_][-1]+1]
-                    }
-                    entities.append(entitie_)
-                    break
+        for entity_ in label_entities:
+            entities.append({
+                "start_idx": entity_[1],
+                "end_idx": entity_[2],
+                "entity": text[entity_[1]: entity_[2]+1],
+                "type": entity_[0]
+            })
 
         return entities
