+++ ark_nlp/model/ner/global_pointer_bert/__init__.py
@@ -1,5 +1,5 @@
-from ark_nlp.dataset import GlobalPointerNERDataset as Dataset 
-from ark_nlp.dataset import GlobalPointerNERDataset as GlobalPointerBertNERDataset
+from ark_nlp.model.ner.global_pointer_bert.dataset import GlobalPointerBertNERDataset as Dataset 
+from ark_nlp.model.ner.global_pointer_bert.dataset import GlobalPointerBertNERDataset
 
 from ark_nlp.processor.tokenizer.transformer import SpanTokenizer as Tokenizer
 from ark_nlp.processor.tokenizer.transformer import SpanTokenizer as GlobalPointerBertNERTokenizer
@@ -7,15 +7,15 @@ from ark_nlp.processor.tokenizer.transformer import SpanTokenizer as GlobalPoint
 from ark_nlp.nn import BertConfig as GlobalPointerBertConfig
 from ark_nlp.nn import BertConfig as ModuleConfig
 
-from ark_nlp.model.ner.global_pointer_bert.global_pointer_bert import GlobalPointerBert
-from ark_nlp.model.ner.global_pointer_bert.global_pointer_bert import EfficientGlobalPointerBert
-from ark_nlp.model.ner.global_pointer_bert.global_pointer_bert import GlobalPointerBert as Module
+from ark_nlp.model.ner.global_pointer_bert.module import GlobalPointerBert
+from ark_nlp.model.ner.global_pointer_bert.module import EfficientGlobalPointerBert
+from ark_nlp.model.ner.global_pointer_bert.module import GlobalPointerBert as Module
 
 from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_model_optimizer
 from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_global_pointer_bert_optimizer
 
-from ark_nlp.model.ner.global_pointer_bert.global_pointer_bert_named_entity_recognition import GlobalPointerNERTask as Task
-from ark_nlp.model.ner.global_pointer_bert.global_pointer_bert_named_entity_recognition import GlobalPointerNERTask as GlobalPointerBertNERTask
+from ark_nlp.model.ner.global_pointer_bert.task import GlobalPointerBertNERTask as Task
+from ark_nlp.model.ner.global_pointer_bert.task import GlobalPointerBertNERTask
 
-from ark_nlp.factory.predictor import GlobalPointerNERPredictor as Predictor
-from ark_nlp.factory.predictor import GlobalPointerNERPredictor as GlobalPointerBertNERPredictor
+from ark_nlp.model.ner.global_pointer_bert.predictor import GlobalPointerBertNERPredictor as Predictor
+from ark_nlp.model.ner.global_pointer_bert.predictor import GlobalPointerBertNERPredictor
