+++ ark_nlp/model/ner/global_pointer_bert/dataset.py
@@ -21,9 +21,9 @@ import torch
 from ark_nlp.dataset import TokenClassificationDataset
 
 
-class SpanNERDataset(TokenClassificationDataset):
+class GlobalPointerBertNERDataset(TokenClassificationDataset):
     """
-    用于Span模式的命名实体识别任务的Dataset
+    用于GlobalPointer命名实体识别任务的Dataset
 
     Args:
         data (:obj:`DataFrame` or :obj:`string`): 数据或者数据地址
@@ -35,11 +35,7 @@ class SpanNERDataset(TokenClassificationDataset):
     """  # noqa: ignore flake8"
 
     def _get_categories(self):
-        categories = sorted(list(set([label_['type']
-                                      for data in self.dataset for label_ in data['label']])))
-        if 'O' in categories:
-            categories.remove('O')
-        categories.insert(0, 'O')
+        categories = sorted(list(set([label_['type'] for data in self.dataset for label_ in data['label']])))
         return categories
 
     def _convert_to_transfomer_ids(self, bert_tokenizer):
@@ -56,36 +52,27 @@ class SpanNERDataset(TokenClassificationDataset):
 
             input_ids, input_mask, segment_ids = input_ids
 
-            start_label = torch.zeros((bert_tokenizer.max_seq_len))
+            global_label = torch.zeros((
+                self.class_num,
+                bert_tokenizer.max_seq_len,
+                bert_tokenizer.max_seq_len)
+            )
 
-            end_label = torch.zeros((bert_tokenizer.max_seq_len))
-
-            label_ = set()
             for info_ in row_['label']:
                 if info_['start_idx'] in start_mapping and info_['end_idx'] in end_mapping:
                     start_idx = start_mapping[info_['start_idx']]
                     end_idx = end_mapping[info_['end_idx']]
                     if start_idx > end_idx or info_['entity'] == '':
                         continue
+                    global_label[self.cat2id[info_['type']], start_idx+1, end_idx+1] = 1
 
-                    start_label[start_idx+1] = self.cat2id[info_['type']]
-                    end_label[end_idx+1] = self.cat2id[info_['type']]
-
-                    label_.add((self.cat2id[info_['type']], start_idx, end_idx))
+            global_label = global_label.to_sparse()
 
             features.append({
                 'input_ids': input_ids,
                 'attention_mask': input_mask,
                 'token_type_ids': segment_ids,
-                'start_label_ids': start_label,
-                'end_label_ids': end_label,
-                'label_ids': list(label_)
+                'label_ids': global_label
             })
 
         return features
-
-    @property
-    def to_device_cols(self):
-        _cols = list(self.dataset[0].keys())
-        _cols.remove('label_ids')
-        return _cols
