+++ ark_nlp/model/ner/span_bert/__init__.py
@@ -1,5 +1,5 @@
-from ark_nlp.dataset import SpanNERDataset as Dataset
-from ark_nlp.dataset import SpanNERDataset as SpanBertNERDataset
+from ark_nlp.model.ner.span_bert.dataset import SpanBertNERDataset as Dataset
+from ark_nlp.model.ner.span_bert.dataset import SpanBertNERDataset
 
 from ark_nlp.processor.tokenizer.transformer import TokenTokenizer as Tokenizer
 from ark_nlp.processor.tokenizer.transformer import TokenTokenizer as SpanBertNERTokenizer
@@ -7,14 +7,14 @@ from ark_nlp.processor.tokenizer.transformer import TokenTokenizer as SpanBertNE
 from ark_nlp.nn import BertConfig as SpanBertConfig
 from ark_nlp.nn import BertConfig as ModuleConfig
 
-from ark_nlp.model.ner.span_bert.span_bert import SpanIndependenceBert as SpanBert
-from ark_nlp.model.ner.span_bert.span_bert import SpanIndependenceBert as Module
+from ark_nlp.model.ner.span_bert.module import SpanIndependenceBert as SpanBert
+from ark_nlp.model.ner.span_bert.module import SpanIndependenceBert as Module
 
 from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_model_optimizer
 from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_span_bert_optimizer
 
-from ark_nlp.model.ner.span_bert.span_bert_named_entity_recognition import SpanNERTask as Task
-from ark_nlp.model.ner.span_bert.span_bert_named_entity_recognition import SpanNERTask as SpanBertNERTask
+from ark_nlp.model.ner.span_bert.task import SpanBertNERTask as Task
+from ark_nlp.model.ner.span_bert.task import SpanBertNERTask
 
-from ark_nlp.factory.predictor import SpanNERPredictor as Predictor
-from ark_nlp.factory.predictor import SpanNERPredictor as SpanBertNERPredictor
+from ark_nlp.model.ner.span_bert.predictor import SpanBertNERPredictor as Predictor
+from ark_nlp.model.ner.span_bert.predictor import SpanBertNERPredictor
