+++ ark_nlp/model/ner/w2ner_bert/__init__.py
@@ -1,4 +1,5 @@
-from ark_nlp.model.ner.w2ner_bert.w2ner_named_entity_recognition_dataset import W2NERDataset as Dataset
+from ark_nlp.model.ner.w2ner_bert.dataset import W2NERDataset as Dataset
+from ark_nlp.model.ner.w2ner_bert.dataset import W2NERDataset
 
 from ark_nlp.processor.tokenizer.transformer import TokenTokenizer as Tokenizer
 from ark_nlp.processor.tokenizer.transformer import TokenTokenizer as W2NERTokenizer
@@ -6,13 +7,14 @@ from ark_nlp.processor.tokenizer.transformer import TokenTokenizer as W2NERToken
 from ark_nlp.nn import BertConfig as W2NERBertConfig
 from ark_nlp.nn import BertConfig as ModuleConfig
 
-from ark_nlp.model.ner.w2ner_bert.w2ner_bert import W2NERBert
-from ark_nlp.model.ner.w2ner_bert.w2ner_bert import W2NERBert as Module
+from ark_nlp.model.ner.w2ner_bert.module import W2NERBert
+from ark_nlp.model.ner.w2ner_bert.module import W2NERBert as Module
 
 from ark_nlp.factory.optimizer import get_w2ner_model_optimizer as get_default_model_optimizer
 from ark_nlp.factory.optimizer import get_w2ner_model_optimizer as get_default_w2ner_optimizer
 
-from ark_nlp.model.ner.w2ner_bert.w2ner_named_entity_recognition_task import W2NERTask as Task
+from ark_nlp.model.ner.w2ner_bert.task import W2NERTask as Task
+from ark_nlp.model.ner.w2ner_bert.task import W2NERTask
 
-from ark_nlp.model.ner.w2ner_bert.w2ner_named_entity_recognition_predictor import W2NERPredictor as Predictor
-from ark_nlp.model.ner.w2ner_bert.w2ner_named_entity_recognition_predictor import W2NERPredictor
+from ark_nlp.model.ner.w2ner_bert.predictor import W2NERPredictor as Predictor
+from ark_nlp.model.ner.w2ner_bert.predictor import W2NERPredictor
