+++ ark_nlp/model/prompt/prompt_ernie_ctm_nptag/module.py
