+++ ark_nlp/model/rc/pure_bert/__init__.py
@@ -1,5 +1,5 @@
-from ark_nlp.model.rc.pure_bert.pure_relation_classification_dataset import PURERCDataset
-from ark_nlp.model.rc.pure_bert.pure_relation_classification_dataset import PURERCDataset as Dataset
+from ark_nlp.model.rc.pure_bert.dataset import PURERCDataset
+from ark_nlp.model.rc.pure_bert.dataset import PURERCDataset as Dataset
 
 from ark_nlp.processor.tokenizer.transformer import SpanTokenizer as Tokenizer
 from ark_nlp.processor.tokenizer.transformer import SpanTokenizer as PURERETokenizer
@@ -7,14 +7,14 @@ from ark_nlp.processor.tokenizer.transformer import SpanTokenizer as PUREREToken
 from ark_nlp.nn import BertConfig as PUREBertConfig
 from ark_nlp.nn import BertConfig as ModuleConfig
 
-from ark_nlp.model.rc.pure_bert.pure_bert import PUREBert
-from ark_nlp.model.rc.pure_bert.pure_bert import PUREBert as Module
+from ark_nlp.model.rc.pure_bert.module import PUREBert
+from ark_nlp.model.rc.pure_bert.module import PUREBert as Module
 
 from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_model_optimizer
 from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_pure_bert_optimizer
 
-from ark_nlp.model.rc.pure_bert.pure_relation_classification_task import PURERCTask as Task
-from ark_nlp.model.rc.pure_bert.pure_relation_classification_task import PURERCTask as PURERCTask
+from ark_nlp.model.rc.pure_bert.task import PURERCTask as Task
+from ark_nlp.model.rc.pure_bert.task import PURERCTask as PURERCTask
 
-from ark_nlp.model.rc.pure_bert.pure_relation_classification_predictor import PURERCPredictor as Predictor
-from ark_nlp.model.rc.pure_bert.pure_relation_classification_predictor import PURERCPredictor as PURERCPredictor
+from ark_nlp.model.rc.pure_bert.predictor import PURERCPredictor as Predictor
+from ark_nlp.model.rc.pure_bert.predictor import PURERCPredictor as PURERCPredictor
