+++ ark_nlp/model/re/casrel_bert/__init__.py
@@ -1,5 +1,5 @@
-from ark_nlp.model.re.casrel_bert.casrel_relation_extraction_dataset import CasRelREDataset 
-from ark_nlp.model.re.casrel_bert.casrel_relation_extraction_dataset import CasRelREDataset as Dataset
+from ark_nlp.model.re.casrel_bert.dataset import CasRelREDataset 
+from ark_nlp.model.re.casrel_bert.dataset import CasRelREDataset as Dataset
 
 from ark_nlp.processor.tokenizer.transformer import SpanTokenizer as Tokenizer
 from ark_nlp.processor.tokenizer.transformer import SpanTokenizer as CasRelRETokenizer
@@ -7,14 +7,14 @@ from ark_nlp.processor.tokenizer.transformer import SpanTokenizer as CasRelRETok
 from ark_nlp.nn import BertConfig as CasRelBertConfig
 from ark_nlp.nn import BertConfig as ModuleConfig
 
-from ark_nlp.model.re.casrel_bert.casrel_bert import CasRelBert
-from ark_nlp.model.re.casrel_bert.casrel_bert import CasRelBert as Module
+from ark_nlp.model.re.casrel_bert.module import CasRelBert
+from ark_nlp.model.re.casrel_bert.module import CasRelBert as Module
 
 from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_model_optimizer
 from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_casrel_bert_optimizer
 
-from ark_nlp.model.re.casrel_bert.casrel_relation_extraction_task import CasRelRETask as Task
-from ark_nlp.model.re.casrel_bert.casrel_relation_extraction_task import CasRelRETask as CasRelRETask
+from ark_nlp.model.re.casrel_bert.task  import CasRelRETask as Task
+from ark_nlp.model.re.casrel_bert.task  import CasRelRETask
 
-from ark_nlp.model.re.casrel_bert.casrel_relation_extraction_predictor import CasRelREPredictor as Predictor
-from ark_nlp.model.re.casrel_bert.casrel_relation_extraction_predictor import CasRelREPredictor as CasRelREPredictor
+from ark_nlp.model.re.casrel_bert.predictor  import CasRelREPredictor as Predictor
+from ark_nlp.model.re.casrel_bert.predictor  import CasRelREPredictor
