+++ ark_nlp/model/re/prgc_bert/__init__.py
@@ -1,5 +1,5 @@
-from ark_nlp.model.re.prgc_bert.prgc_relation_extraction_dataset import PRGCREDataset 
-from ark_nlp.model.re.prgc_bert.prgc_relation_extraction_dataset import PRGCREDataset as Dataset
+from ark_nlp.model.re.prgc_bert.dataset import PRGCREDataset 
+from ark_nlp.model.re.prgc_bert.dataset import PRGCREDataset as Dataset
 
 from ark_nlp.processor.tokenizer.transformer import SpanTokenizer as Tokenizer
 from ark_nlp.processor.tokenizer.transformer import SpanTokenizer as PRGCRETokenizer
@@ -7,14 +7,14 @@ from ark_nlp.processor.tokenizer.transformer import SpanTokenizer as PRGCREToken
 from ark_nlp.nn import BertConfig as PRGCBertConfig
 from ark_nlp.nn import BertConfig as ModuleConfig
 
-from ark_nlp.model.re.prgc_bert.prgc_bert import PRGCBert
-from ark_nlp.model.re.prgc_bert.prgc_bert import PRGCBert as Module
+from ark_nlp.model.re.prgc_bert.module import PRGCBert
+from ark_nlp.model.re.prgc_bert.module import PRGCBert as Module
 
 from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_model_optimizer
 from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_prgc_bert_optimizer
 
-from ark_nlp.model.re.prgc_bert.prgc_relation_extraction_task import PRGCRETask as Task
-from ark_nlp.model.re.prgc_bert.prgc_relation_extraction_task import PRGCRETask as PRGCRETask
+from ark_nlp.model.re.prgc_bert.task import PRGCRETask as Task
+from ark_nlp.model.re.prgc_bert.task import PRGCRETask
 
-from ark_nlp.model.re.prgc_bert.prgc_relation_extraction_predictor import PRGCREPredictor as Predictor
-from ark_nlp.model.re.prgc_bert.prgc_relation_extraction_predictor import PRGCREPredictor as PRGCREPredictor
+from ark_nlp.model.re.prgc_bert.predictor import PRGCREPredictor as Predictor
+from ark_nlp.model.re.prgc_bert.predictor import PRGCREPredictor
